requirements.txt
fastapi
pydantic
SQLAlchemy
sqlalchemy_utils
sqlmodel
uvicorn
python-dotenv
psycopg2
passlib
pyjwt
datetime
bcrypt
sqlalchemy-cockroachdb
boto3
python-multipart
python-dotenv